build: enforce strictest C89 compliance and fix CI compilation errors
- Enforced -Wconversion, -Wsign-conversion, -Wshorten-64-to-32, -Wmissing-prototypes and equivalent strict flags in CMakeLists.txt.
- Integrated MSVC Safe CRT functions conditionally via _MSC_VER to resolve C4996 errors.
- Fixed implicit signedness and integer precision loss conversions.
- Resolved missing prototype errors.
- Refactored WebSocket internal frames to correctly handle opcodes.
- Formatted with clang-format.

+++ src/event_loop.c
@@ -768,7 +768,7 @@ enum c_abstract_http_error http_loop_run(struct ModalityEventLoop *loop) {
 
     if (next_timeout >= 0) {
       tv.tv_sec = (long)(next_timeout / 1000);
-      tv.tv_usec = (long)((next_timeout % 1000) * 1000);
+      tv.tv_usec = (int)((next_timeout % 1000) * 1000);
       ptv = &tv;
     }
